requirements.txt
Django==5.2.4
django-allauth==65.15.1
gunicorn==21.2.0
whitenoise==6.8.2
psycopg2-binary==2.9.10
python-decouple==3.8
dj-database-url==2.3.0
Pillow==12.1.1
oauthlib==3.3.1
requests==2.33.1
cloudinary==1.42.0
django-cloudinary-storage==0.3.0
